// PortalAutoservicio/src/app/models/concurso-escalafonario-vigente.ts
export class ConcursoEscalafonarioVigente {

    NUMERO: string;
    NOMBREPUESTO: string;
    JORNADALABORAL: string;
    NIVEL: string;
    RANGO: string;
    NOPLAZA: string;
    PERCEPCIONMENSUALES: string;
    ADSCRIPCION: string;
    UBICACIONTRABAJO: string;
    PUNTAJEESCALAFONARIOMINIMO: string;
    ESCOLARIRADMINIMA: string;
    ESCOLARIRADMINIMAGEN: string;
    EXPERIENCIAMINIMA: string;
    CONOCIMIENTOSMINIMOS: string;
    IDESTATUSPLAZA: string;
    RUTAGUIAESTUDIO: string;
    IDPUESTO: string;
    IDESCOLARIDADMINIMAGEN: string;
    IDESCOLARIDADMINIMAGNAT: string;
    PUESTOINFERIOR: string;
    /*idPlaza:string;
    idEscolaridadMinimaGen:string;
    idEscolaridadMinimaNat:string;
    rutaGuiaEstudio:string;*/
    PVISIBLE: string;
    IDCOLORSEMAFORO: number;
    COLORSEMAFORO: string;
    TOOLTIP: string;
}